requirements.txt
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
openai==1.51.0
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
pandas==2.2.2
openpyxl==3.1.5
tabulate==0.9.0
